fix: preserve embedding outage status for trigger endpoints
